+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Compatibility:
 * Implemented `Float#{floor, ceil}` with `ndigits` argument.
 * Implemented `Thread#fetch`.
 * Implemented `Float#truncate` with `ndigits` argument.
+* Made `String#{byteslice, slice, slice!}` and `Symbol#slice` compatible with endless ranges.
 
 * Implemented `-p` CLI option.
 * Implemented "instance variable not initialized" warning.